fix: :zap:  APKs now wokring correctly!
Since commit [e50b2ac](https://github.com/YeyoM/wellness-mobile-app/commit/e50b2acbabd23fba346e5c57d0766381cda67be0) on May 12th where I upgraded expo to version ~51.0.2 is where the app started crashing, but we did not noticed untill a couple of days ago, where one of our Members noticed that the las aab artifact that we published to our closed alpha realase was corrupted in some form, when opening the app, this just crashed a second after opening.

The initial thought of what might have happened was that the version of expo itself was causing the problem, but further investigating we are in the current state...

We can firmly confirm that everything before the commit mentioned before was working correctly, therefore, the commit in question is where things started going wrong.

Here are some possible options and hypothesis on what is going wrong and crashing the app, and how we investigated in order to find the possible root of the problems...

The first step we took in order to find the possible problem was to try and check the logs on when the app crashed, so after investigationg how to get this logs we finally were able to get those and here is a really small part of the file the bug report shows (the error logs starts at lines 44100 aprox):

```txt

...

05-27 15:47:40.649 10190  4124  4241 E ReactNativeJS: Invariant Violation: requireNativeComponent: "RNSScreenStackHeaderConfig" was not found in the UIManager.
...
05-27 15:47:40.671 10190  4124  4241 E ReactNativeJS: Invariant Violation: requireNativeComponent: "RNSScreenStackHeaderConfig" was not found in the UIManager.
...
05-27 15:47:40.749 10190  4124  4242 E unknown:ReactNative: Invariant Violation: requireNativeComponent: "RNSScreenStackHeaderConfig" was not found in the UIManager.
...
05-27 15:47:40.769 10127  2361  3727 I BugleRcsProvisioning: RcsProvisioningTrigger: onRcsAvailabilityUpdate : availability is updated to Carrier RCS is not set up, lastAvailability is Carrier RCS is not set up, simId hash = 6574909045491450823, simId = redacted-pii:sim_id[chars:20,last3:897], isPev3Enabled = true
...
--------- beginning of crash
05-27 15:47:40.769 10190  4124  4242 E AndroidRuntime: FATAL EXCEPTION: mqt_native_modules
05-27 15:47:40.769 10190  4124  4242 E AndroidRuntime: Process: com.yeyom.wellness, PID: 4124
05-27 15:47:40.769 10190  4124  4242 E AndroidRuntime: com.facebook.react.common.JavascriptException: Invariant Violation: requireNativeComponent: "RNSScreenStackHeaderConfig" was not found in the UIManager.
...

```

here we can check the error is pretty clear, there is something in the build that is missing, by investigating in some forums, we find that the error might be corrected by doing installing some libraries:

react-native-screens react-native-safe-area-context

The above was extracted from... https://stackoverflow.com/questions/69043806/requirenativecomponent-rnsscreenstackheaderconfig-was-not-found-in-the-uimana

And what we found when comparing the package.json files from the last version that was working correctly to the one when we upgraded expo to the new version we found that those libraries were not installed, you can check the full diff on the commit were the issue was [e50b2ac](https://github.com/YeyoM/wellness-mobile-app/commit/e50b2acbabd23fba346e5c57d0766381cda67be0).

Therefore, we are adding the following dependencies to the current package.json:

```json
react-dom,
react-native-safe-area-context,
react-native-screens,
react-native-tab-view,
react-native-web,
react-native-svg,
expo-sharing
```

There are some dependencies from above that might not solve the issue itself, but since they were on the last version that still worked, we are adding them.

This apparently solved the first issue, but now, another issue has appeared, and here are the error logs...

```txt

05-28 11:45:52.506  1000   577   896 I am_crash: [13933,0,com.yeyom.wellness,547929668,com.facebook.react.common.JavascriptException,TypeError: Cannot read property 'onAuthStateChanged' of undefined
05-28 11:45:52.506  1000   577   896 I am_crash:
05-28 11:45:52.506  1000   577   896 I am_crash: This error is located at:
05-28 11:45:52.506  1000   577   896 I am_crash:     in CreateRoutineProvider
05-28 11:45:52.506  1000   577   896 I am_crash:     in InitialScreensProvider
05-28 11:45:52.506  1000   577   896 I am_crash:     in AppContextProvider
05-28 11:45:52.506  1000   577   896 I am_crash:     in App
05-28 11:45:52.506  1000   577   896 I am_crash:     in RCTView
05-28 11:45:52.506  1000   577   896 I am_crash:     in Unknown
05-28 11:45:52.506  1000   577   896 I am_crash:     in AppContainer, js engine: hermes, stack:
05-28 11:45:52.506  1000   577   896 I am_crash: anonymous@1:2028895
05-28 11:45:52.506  1000   577   896 I am_crash: commitHookEffectListMount@1:433396
05-28 11:45:52.506  1000   577   896 I am_crash: commitHookPassiveMountEffects@1:441154
05-28 11:45:52.506  1000   577   896 I am_crash: commitPassiveMountOnFiber@1:441604
05-28 11:45:52.506  1000   577   896 I am_crash: recursivelyTraversePassiveMountEffects@1:441239
05-28 11:45:52.506  1000   577   896 I am_crash: commitPassiveMountOnFiber@1:441347
05-28 11:45:52.506  1000   577   896 I am_crash: recursivelyTraversePassiveMountEffects@1:441239
05-28 11:45:52.506  1000   577   896 I am_crash: commitPassiveMountOnFiber@1:441577
05-28 11:45:52.506  1000   577   896 I am_crash: recursivelyTraversePassiveMountEffects@1:441239
05-28 11:45:52.506  1000   577   896 I am_crash: commitPassiveMountOnFiber@1:441347
05-28 11:45:52.506  1000   577   896 I am_crash: recursivelyTraversePassiveMountEffects@1:441239
05-28 11:45:52.506  1000   577   896 I am_crash: commitPassiveMountOnFiber@1:441577
05-28 11:45:52.506  1000   577   896 I am_crash: recursivelyTraversePassiveMountEffects@1:441239
05-28 11:45:52.506  1000   577   896 I am_crash: commitPassiveMountOnFiber@1:441577
05-28 11:45:52.506  1000   577   896 I am_crash: recursivelyTraversePassiveMountEffects@1:441239
05-28 11:45:52.506  1000   577   896 I am_crash: commitPassiveMountOnFiber@1:441347
05-28 11:45:52.506  1000   577   896 I am_crash: recursivelyTraversePassiveMountEffects@1:4412^
05-28 11:45:52.515  1000   577   896 I wm_finish_activity: [0,81570307,43,com.yeyom.wellness/.MainActivity,force-crash]

```